Testing cross-platform malloc_func_type usage
GitHub's tests on a Windows target have been failing, apparently because of some quirk with how ``typeid`` or ``std::type_index`` assigns values. They should be the same between a function pointer "type" (an instantiation of ``std::function``) and a function that's compatible with it. On Windows, the values are slightly different. This is okay as far as background implementation details go, so long as ``malloc_func_type`` can simply be assigned to ``std::malloc``.